refactor: move functions check_dir and make_dir to utils.py
